requirements.txt
streamlit==1.29.0
ollama==0.1.6
Pillow==10.1.0 